Chapter7_SportsStore/SportsStore.WebUI/Controllers/AccountController.cs
using SportsStore.WebUI.Infrastructure.Abstract;
using SportsStore.WebUI.Models;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;
using System.Web.Mvc;

namespace SportsStore.WebUI.Controllers
{
    public class AccountController : Controller
    {
        private readonly IAuthProvider _authProvider;

        public AccountController(IAuthProvider authProvider)
        {
            _authProvider = authProvider;
        }

        // GET: AccountController
        public ActionResult Login()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Login(LoginViewModel model, string returnUrl)
        {
            if (ModelState.IsValid)
            {
                var authenticated = _authProvider.Authenticate(model.UserName, model.Password);

                if (authenticated)
                {
                    if (string.IsNullOrEmpty(returnUrl))
                    {
                        return RedirectToAction("Index", "Admin");
                    }
                    else {
                        return Redirect(returnUrl);
                     }                   
                }
                ModelState.AddModelError("", "Incorrect username or password");
            }

            return View(model);
        }
    }
}